refactor: убраны лишние аннотации column, for(i) заменен на foreach fix: переписано решение рейтингов

+++ ewm-main/src/main/java/ru/practicum/dto/EventFullDto.java
@@ -50,5 +50,5 @@ public class EventFullDto {
 
     private Long views;
 
-    private double rating;
+    private Long rating;
 }
