+++ ewm-main/src/main/java/ru/practicum/dto/UserDto.java
@@ -25,6 +25,6 @@ public class UserDto {
     @NotBlank
     private String name;
 
-    private double rating;
+    private Long rating;
 
 }
