+++ ewm-main/src/main/java/ru/practicum/dto/UserShortDto.java
@@ -12,5 +12,5 @@ import lombok.RequiredArgsConstructor;
 public class UserShortDto {
     private Long id;
     private String name;
-    private double rating;
+    private Long rating;
 }
